Organize entity scripts
