+++ frontend/src/entities/arrow.js
@@ -2,8 +2,8 @@ export const ARROW_SPEED = 3;
 export const ARROW_DAMAGE = 2;
 export const ARROW_PREVIEW_RANGE = 12 * 40;
 
-import { circleRectColliding, isColliding } from "./game_logic.js";
-import { damageWall } from "./walls.js";
+import { circleRectColliding, isColliding } from "../game_logic.js";
+import { damageWall } from "../walls.js";
 
 export function predictArrowEndpoint(x, y, direction, walls, zombies = []) {
   const len = Math.hypot(direction.x, direction.y);
