+++ frontend/tests/orbs.test.js
@@ -1,6 +1,6 @@
 import test from "node:test";
 import assert from "node:assert/strict";
-import { createOrbs, updateOrbs } from "../src/orbs.js";
+import { createOrbs, updateOrbs } from "../src/entities/orbs.js";
 
 // stub for collision
 import { isColliding } from "../src/game_logic.js";
